+++ MaquetaParaFinal/Clases/Personal.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using MaquetaParaFinal.View.Agregar;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -20,7 +21,8 @@ namespace MaquetaParaFinal.View
 
         private void btAgregar_Click(object sender, RoutedEventArgs e)
         {
-            //TO-DO
+            AgregarPersonal ap = new AgregarPersonal();
+            ap.ShowDialog();
         }
 
         private void btModificar_Click(object sender, RoutedEventArgs e)
